Create Syllabus controller
Moved syllabus controller and associated method to its own controller
file.

Created syllabus controller spec and moved associated specs for
controller syllabus to its own file.

Updated syllabus controller route. Syllabus action name is now 'update'.
Routing endpoint remains the same.

This helps to keep controller's small and manageable.
