+++ config/routes.rb
@@ -100,7 +100,7 @@ Rails.application.routes.draw do
           titleterm: /[^\/]*/
         }
     post 'clone_course/:id' => 'course_clone#clone'
-    post 'courses/:id/update_syllabus' => 'courses#update_syllabus'
+    post 'courses/:id/update_syllabus' => 'courses/syllabuses#update'
     delete 'courses/:id/delete_all_weeks' => 'courses#delete_all_weeks',
       constraints: {
         id: /.*/
